refactor: presenter removed and andlers created
