+++ usecase/interfaces/character_repository.go
