--- usecases/interfaces/get_character_presenter.go
@@ -1,9 +1,0 @@
-package interfaces
-
-import (
-	"bootcamp/domain/model"
-)
-
-type GetCharacterPresenter interface {
-	Present(character *model.Character) string
-}
